Exract out flattening s-expressions into an integer sequence.
