+++ Makefile
@@ -94,6 +94,7 @@ SEXP_OBJDIR = $(OBJDIR)/sexp
 SEXP_OBJDIR_BOOT = $(OBJDIR_BOOT)/sexp
 SEXP_SRCS_BASE = \
 	Ast.cpp \
+	AstToCasm.cpp \
 	TextWriter.cpp \
 	TraceSexp.cpp
