+++ src/interp/IntFormats.h
@@ -47,10 +47,6 @@ enum class IntTypeFormat {
 
 const char* getName(IntTypeFormat Fmt);
 
-extern bool definesIntTypeFormat(const filt::Node* Nd);
-extern bool definesLitIntTypeFormat(const filt::Node* Nd);
-extern IntTypeFormat getIntTypeFormat(const filt::Node* Nd);
-
 static constexpr size_t NumIntTypeFormats = size_t(IntTypeFormat::LAST) + 1;
 
 class IntTypeFormats {
